+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 + [x] Table: updated for using with Datetime (change `toDate()` to `Datetime.from()`)
 + [x] List: updated for using with Datetime (change `toDate()` to `Datetime.from()`)
 + [x] Datepicker: updated for using with Datetime
++ [x] Utilities: updated for using with Datetime
 
 ### 4.4.3
 #### M4Q